Implements drawer and header navigation
Adds navigation functionality to the drawer and stack header.

The header left icon now toggles the drawer if the current screen is the initial stack screen, otherwise, it acts as a back button.  A button to open the drawer has also been added to the home screen for testing purposes.  Header and tab styling has been adjusted for better visibility.